+++ src/Router.js
@@ -1,9 +1,8 @@
-import React from 'react';
 import { BrowserRouter, Routes, Route } from 'react-router-dom';
 import Nav from './components/Nav/Nav';
 import Main from './pages/Main/Main';
 import Login from './pages/Login/Login';
-import SignUp2 from './pages/SignUp2/SignUp2';
+import SignUp from './pages/SignUp/SignUp';
 import Footer from './components/Footer/Footer';
 
 const Router = () => {
@@ -13,7 +12,7 @@ const Router = () => {
       <Routes>
         <Route path="/" element={<Main />} />
         <Route path="login" element={<Login />} />
-        <Route path="signup" element={<SignUp2 />} />
+        <Route path="signup" element={<SignUp />} />
       </Routes>
       <Footer />
     </BrowserRouter>
