--- src/pages/SignUp/components/SignUpInputList.js
@@ -1,44 +1,0 @@
-// export const SIGNUPINPUT_LIST = [
-//   {
-//     id: 1,
-//     title: '이메일',
-//     type: 'email',
-//     placeholder: '이메일을 입력해주세요.',
-//     isThisBtn: true,
-//   },
-//   {
-//     id: 2,
-//     title: '비밀번호',
-//     type: 'password',
-//     placeholder: '비밀번호를 입력해주세요.',
-//     isThisBtn: false,
-//   },
-//   {
-//     id: 3,
-//     title: '비밀번호확인',
-//     type: 'password',
-//     placeholder: '비밀번호를 한 번 더 입력해주세요.',
-//     isThisBtn: false,
-//   },
-//   {
-//     id: 4,
-//     title: '이름',
-//     type: 'text',
-//     placeholder: '이름을 입력해주세요.',
-//     isThisBtn: false,
-//   },
-//   {
-//     id: 5,
-//     title: '휴대폰',
-//     type: 'tel',
-//     placeholder: '이름을 입력해주세요.',
-//     isThisBtn: false,
-//   },
-//   {
-//     id: 6,
-//     title: '주소',
-//     type: 'text',
-//     placeholder: '주소를 입력해주세요.',
-//     isThisBtn: false,
-//   },
-// ];
