+++ src/pages/SignUp/components/signUpItem/SignUpItem.scss
@@ -3,8 +3,8 @@
 .signUpItem {
   display: grid;
   grid-template-columns: 120px 333px 120px;
-  width: 640px;
   grid-column-gap: 10px;
+  width: 640px;
   padding: 10px 20px;
 
   .nameSection {
@@ -20,9 +20,9 @@
 
   .comfirmRepetition {
     background-color: white;
-    color: $basic-color;
     border: 1px solid $basic-color;
     border-radius: 4px;
+    color: $basic-color;
     cursor: pointer;
   }
 }
