+++ src/pages/SignUp/components/userAgreement/UserAgreement.scss
@@ -1,0 +1,44 @@
+@import './../../../../styles/variables.scss';
+
+.userAgreement {
+  display: flex;
+  width: 640px;
+  padding-top: 20px;
+
+  .userAgreementTitle {
+    width: 139px;
+    height: 260px;
+    padding-left: 20px;
+  }
+  .allAgreementTop {
+    display: flex;
+    align-items: center;
+
+    .allAgreementPhrase {
+      text-indent: 10px;
+      font-size: 18px;
+    }
+  }
+
+  .checkedIcon,
+  .uncheckedIcon {
+    width: 24px;
+    height: 24px;
+    cursor: pointer;
+  }
+
+  .checkedIcon {
+    color: $basic-color;
+  }
+
+  .uncheckedIcon {
+    color: $border-color;
+  }
+
+  .disagreementPhrase {
+    margin-top: 5px;
+    text-indent: 35px;
+    font-size: 12px;
+    color: $border-color;
+  }
+}
