+++ src/pages/SignUp/components/userAgreement/UserAgreementCheck.scss
@@ -1,0 +1,29 @@
+@import './../../../../styles/variables.scss';
+
+.UserAgreementCheck {
+  @include flexRow(space-between, center);
+
+  .agreementCheckSection {
+    display: flex;
+    align-items: center;
+    margin-top: 20px;
+
+    .agreementContent {
+      text-indent: 10px;
+    }
+
+    .mustOrChoice {
+      padding-left: 5px;
+      color: $border-color;
+    }
+  }
+
+  .termsAndConditionsBtn {
+    padding: 8px 10px 0 0;
+    background-color: white;
+    border: none;
+    color: $basic-color;
+    font-size: 14px;
+    cursor: pointer;
+  }
+}
