+++ src/pages/SignUp/components/userAgreement/UserAgreementList.js
@@ -1,0 +1,26 @@
+export const USERAGREEMENT_LIST = [
+  {
+    id: 1,
+    content: '이용약관 동의',
+    option: '필수',
+    button: true,
+  },
+  {
+    id: 2,
+    content: '개인정보 수집・이용 동의',
+    option: '필수',
+    button: true,
+  },
+  {
+    id: 3,
+    content: '개인정보 수집・이용 동의',
+    option: '선택',
+    button: true,
+  },
+  {
+    id: 1,
+    content: '본인은 만 14세 이상입니다.',
+    option: '필수',
+    button: false,
+  },
+];
