--- src/pages/SignUp2/components/userAgreement/UserAgreement.scss
@@ -1,15 +1,0 @@
-.userAgreement {
-  display: flex;
-  width: 640px;
-  padding-top: 10px;
-
-  .userAgreementTitle {
-    width: 139px;
-    height: 330px;
-    padding-left: 20px;
-  }
-
-  .allAgreementPhrase {
-    font-size: 18px;
-  }
-}
